Move BindingType from the library to the tests, the only use of the class.
