--- Runtime/Bindings/Reflection/BindingType.cs
@@ -1,26 +1,0 @@
-﻿using System;
-
-namespace TechTalk.SpecFlow.Bindings.Reflection
-{
-    public class BindingType : IBindingType
-    {
-        public string Name { get; private set; }
-        public string FullName { get; private set; }
-
-        public Type Type { get { throw new NotImplementedException (); } }
-
-        public BindingType(string name, string fullName)
-        {
-            if (name == null) throw new ArgumentNullException("name");
-            if (fullName == null) throw new ArgumentNullException("fullName");
-
-            Name = name;
-            FullName = fullName;
-        }
-
-        public override string ToString()
-        {
-            return "[" + FullName + "]";
-        }
-    }
-}
